build: add nix flake, derivation, and shell
Add nix build files.

This commit is added for my convenience, because I like to build
software with nix while I work on it. It's not targeting upstream,
though I may try to add the project as a package in nixpkgs.

See https://nixos.org/guides/nix-pills/01-why-you-should-give-it-a-try
and https://github.com/NixOS/nixpkgs